werkflow-docker/pyproject.toml
[project]
name = "werkflow-docker"
version = "0.3.0"
description = "An Docker module for Werkflow."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Ada Lundhe", email = "adalundhe@lundhe.audio"}
]
classifiers = [
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Operating System :: OS Independent",
]
keywords = [
    'pypi', 
    'cicd', 
    'python',
    'werkflow',
    'workflow',
    'docker'
]
license-files = [ "LICENSE" ]
dependencies = [
    'werkflow-core',
    'docker',
    'werkflow-file',
    'werkflow-secrets',
    'werkflow-shell',
    'werkflow-system',
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"